+++ app.py
@@ -5,8 +5,8 @@
 from flask_migrate import Migrate, MigrateCommand
 from flask_script import Manager
 
-from models import *
-from settings.config import create_app
+from apps.models import *
+from settings.config import create_app, db
 
 app = create_app(config="dev_config.py")
