+++ apps/utils/FormatStr.py
@@ -1,0 +1,6 @@
+#!/usr/bin/env python
+# -*- coding: utf-8 -*-
+# @Time    : 2019/8/5 15:55
+# @File    : FormatStr.py
+# @author  : dfkai
+# @Software: PyCharm
