+++ apps/utils/creaet_model/createModels.py
@@ -3,7 +3,7 @@ import os
 
 data = """
 # -*- coding: utf-8 -*-
-from models import db,CRUDMixin
+from settings.dataBase import CRUDMixin, db
 
 
 class {model_Name}(db.Model,CRUDMixin):
