+++ settings/dev_config.py
@@ -15,6 +15,6 @@ delta = datetime.timedelta(hours=10)
 DEBUG = True
 
 # db
-# SQLALCHEMY_DATABASE_URI = f'sqlite:///{dev_db}'
-SQLALCHEMY_DATABASE_URI = 'mysql+pymysql://root:123456@127.0.0.1:3306/dev?charset=utf8'
+SQLALCHEMY_DATABASE_URI = f'sqlite:///{dev_db}'
+# SQLALCHEMY_DATABASE_URI = 'mysql+pymysql://root:123456@127.0.0.1:3306/dev?charset=utf8'
 SQLALCHEMY_TRACK_MODIFICATIONS = True
